Moved logic to a separate function
More details on this change that I made:

There was a file with some looping and printing in the main function.
I moved all the business logic to a separtae function and called that function from the main function.

Why?
This adds organization to the C code, makes future changes easier etc. etc. All the bullshit reasons you wanna give for any change you want to make :D so the project manager knows what is going on...